LayoutTests/fast/events/hittest-pointer-event-none-scrollbars-expected.txt

PASS verticalScrollbarHit is document.body
PASS horizontalScrollbarHit is document.body
PASS resizerHit is document.body
PASS successfullyParsed is true

TEST COMPLETE
